Rewrite the rcS boot script, this time for real!
This simplifies a lot of things, including the autologin mechanism (that
now does not rely on the root's .profile anymore).

Still to be done: check if everything still work (in particular, if
partitions are correctly identified and mounted), and factorise the
partition mount mechanism (after making sure that it does not break).
It can also be simplified a lot...