Add a first version of the script to build a NENA-enabled software router
This allows to build a host image (without VM and guest) to be used
as a software router. The script also allows to install a Preempt-RT
kernel, and the NENA driver for e1000e

NENA (Never Ever NAPI Again) is an experimental network architecture
executing all the network stack in the IRQ thread of the NIC. Some details
(especially with multi-queue) still have to be figured out, but the first
prototype seems to be able to eliminate the performance penalty payed by
RT kernels.